Moving Pages to Steps so that we can use dependency injection instead
