+++ tests/_support/Step/Acceptance/DetailView.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace Step\Acceptance;
+
+use \AcceptanceTester as Tester;
+
+
+class DetailView extends Tester
+{
+
+    /**
+     * Click on the action menu and select a menu item
+     * @param string $link
+     * <?php
+     * $detailView = new \Page\DetailView($i);
+     * $detailView->clickActionMenuItem('Edit');
+     *
+     */
+    public function clickActionMenuItem($link)
+    {
+        $I = $this;
+
+        $I->click('ACTIONS', '#tab-actions');
+        $I->waitForElementVisible('#tab-actions > .dropdown-menu');
+
+        $I->click($link, '#tab-actions > .dropdown-menu');
+    }
+}
