+++ tests/_support/Step/Acceptance/ListView.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Step\Acceptance;
+
+use \AcceptanceTester as Tester;
+
+class ListView  extends Tester
+{
+
+    /**
+     *
+     */
+    public function clickFilter()
+    {
+        $I = $this;
+    }
+}
